==> 02-Inference.Rmd <==
---
title: "R Notebook"
output: html_notebook
---
# Import data
```{r}
data <- read.csv("Data/Extended_data_clean.csv", header = T)
```

# Permutational test function
```{r}
###
###Permutation t tests
###

perm_t_test=function(x,y,iter=1e3){
  
  T0=abs(median(x)-median(y))  # define the test statistic
  T_stat=numeric(iter) # a vector to store the values of each iteration
  x_pooled=c(x,y) # pooled sample
  n=length(x_pooled)
  n1=length(x)
  
  for(perm in 1:iter){ # loop for conditional MC
    # permutation:
    permutation <- sample(1:n)
    x_perm <- x_pooled[permutation]
    x1_perm <- x_perm[1:n1]
    x2_perm <- x_perm[(n1+1):n]
    # test statistic:
    T_stat[perm] <- abs(mean(x1_perm) - mean(x2_perm))
    
  }
  
  # p-value
  p_val <- sum(T_stat>=T0)/iter
  return(p_val)
}
```

# Test for public vs private schools
```{r}
###
# Rank test to see if CLASE_TIPO (private vs public) scores are different
###

# x.1 public
# x.2 private

x.1 <- data$P_Puntaje_2021[which(data$CLASE_TIPO == 0)]
x.2 <- data$P_Puntaje_2021[which(data$CLASE_TIPO == 1)]

perm_t_test(x.1, x.2)
```

```{r}
boxplot(
  list("Public" = x.1, "Private" = x.2),
  main = "Scores of public vs private schools",
  xlab = "Type of school",
  ylab = "Score",
  col = c("red", "cyan")
)
```
We can confirm that the private school perform better than the public ones.

# Test for Calendar A vs B
```{r}
###
# Test to see if scores are different for each CALENDARIO
###


x1 <- data$P_Puntaje_2021[which(data$CALENDARIO == 1)]
x2 <- data$P_Puntaje_2021[which(data$CALENDARIO == 3)]

perm_t_test(x1, x2)
```
Schools from Calendario B have different distribution than those from Calendario A.

```{r}
boxplot(list("A" = x1, "B" = x2),
        main = "Scores of calendar A vs B schools",
        xlab = "Type of school",
        ylab = "Score",
        col = c("red", "cyan")
        )

```

With the boxplot we can clearly see that calendar B schools perform significantly better than schools from calendar A.

# Test for city zone (COD_LOCA)
```{r}
###
# Test to see if there are difference in the scores according to the zone (COD_LOCA)
###

table(data$COD_LOCA)
```

```{r}
###
### ANOVA
###

scores <- data$P_Puntaje_2021
LOC <- as.factor(data$COD_LOCA)

fit <- aov(scores ~ LOC)
summary(fit)

T0 <- summary(fit)[[1]][1,4]  # extract the test statistic
```


```{R}
B = 1e3

T_stat <- numeric(B) 
n <- dim(data)[1]

for(perm in 1:B){
  # Permutation:
  permutation <- sample(1:n)
  scores_perm <- scores[permutation]
  fit_perm <- aov(scores_perm ~ LOC)
  
  # Test statistic:
  T_stat[perm] <- summary(fit_perm)[[1]][1,4]
}
```

```{r}
hist(T_stat,xlim=range(c(T_stat,T0)),breaks=30)
abline(v=T0,col=3,lwd=2)
```

```{r}
p_val <- sum(T_stat>=T0)/B
p_val
```

```{r}
g = length(unique(data$COD_LOCA))
plot(scores ~ LOC, xlab='treat', col=rainbow(length(levels(LOC))), main='Original Data')
```

There is statistical evidence to assure that there is a difference in the scores of schools from different zones.

# Test for number of students
```{r}
# see whether the number of students is relevant. 
quantile(data$EVALUADOS_2021,probs=c(0.25,0.5,0.75))

# we use as a threshold 92 students
small_schools = data[which(data$EVALUADOS_2021<=92),]
big_schools = data[which(data$EVALUADOS_2021>92),]

set.seed(123)

perm_t_test=function(x,y,iter=1e3){
  
  T0=abs(median(x)-median(y))  # define the test statistic
  T_stat=numeric(iter) # a vector to store the values of each iteration
  x_pooled=c(x,y) # pooled sample
  n=length(x_pooled)
  n1=length(x)
  
  for(perm in 1:iter){ # loop for conditional MC
    # permutation:
    permutation <- sample(1:n)
    x_perm <- x_pooled[permutation]
    x1_perm <- x_perm[1:n1]
    x2_perm <- x_perm[(n1+1):n]
    # test statistic:
    T_stat[perm] <- abs(median(x1_perm) - median(x2_perm))
    
  }
  
  # p-value
  p_val <- sum(T_stat>=T0)/iter
  return(p_val)
}

B2 = 1e4

perm_t_test(small_schools$P_Puntaje_2021,big_schools$P_Puntaje_2021)
#p-value is 0
```

```{r}
hist(small_schools$P_Puntaje_2021,col = "green")
hist(big_schools$P_Puntaje_2021,col="blue")

#small schools perform better
```

By grouping the schools into small (< quant 75%) and big (> quant 75%) we find a signifincant difference between the scores.


```{r}
boxplot(list("Small" = small_schools$P_Puntaje_2021, "Big" = big_schools$P_Puntaje_2021),
        main = "Scores of small vs big schools",
        xlab = "Type of school",
        ylab = "Score",
        col = c("red", "cyan")
        )

```
