==> 07-Individual_students.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

```{r}
# Import the data
data_bogota <- read.csv("Data/data_students.csv")
```

# ANALYSIS OF PEERS SES

We want to see if the SES of peers is influential when taking out the personal SES effect.
```{r}
# Import the cleaned dataset
data_schools <- read.csv("Data/Extended_data_clean.csv")

# Calculate the mean individual SES for each school and add it as a new column
mean_SES <- c()
for (i in 1:nrow(data_schools)){
  idx <- which(data_bogota$COLE_COD_DANE_SEDE == data_schools$COD_DANE12[i])
  INSE <- data_bogota$ESTU_INSE_INDIVIDUAL[idx]
  INSE <- INSE[which(INSE != "-")]
  SES_i <- na.omit(as.numeric(INSE))
  mean_SES <- c(mean_SES, mean(SES_i))
}

# Add the new column to the schools data set
data_schools <- cbind(data_schools, indiv_SES = mean_SES)
data_schools <- na.omit(data_schools)
```

```{r}
# Add the mean SES value of the school to each student
pb <- txtProgressBar(min = 0,
                     max = nrow(data_bogota),
                     style = 3,
                     width = 50,
                     char = "=")

mean_SES <- c()
for (i in 1:nrow(data_bogota)){
  idx <- which(data_schools$COD_DANE12 == data_bogota$COLE_COD_DANE_SEDE[i])
  SES_i <- data_schools$indiv_SES[idx]
  mean_SES <- c(mean_SES, mean(SES_i))
  setTxtProgressBar(pb, i)
}

df_reg <- cbind(data_bogota[which(!is.na(mean_SES)),],
                mean_ses = mean_SES[which(!is.na(mean_SES))])
df_reg <- na.omit(df_reg)
```


```{r}
# Separate in train/test to evaluate performance
sampled <- sample(c(T,F), nrow(df_reg), replace = T, prob = c(0.7, 0.3))
df_reg_train <- df_reg[sampled,]
df_reg_test <- df_reg[!sampled,]

# Regress the score with mean SES and INDIV_SES
model_full <- lm(PUNT_GLOBAL ~ mean_ses + as.numeric(ESTU_INSE_INDIVIDUAL),
                  data = df_reg_train)
model_peers <- lm(PUNT_GLOBAL ~ mean_ses,
                  data = df_reg_train)
model_indiv <- lm(PUNT_GLOBAL ~ as.numeric(ESTU_INSE_INDIVIDUAL),
                  data = df_reg_train)

summary(model_full)
summary(model_peers)
summary(model_indiv)
```

```{r}
# Comparison of the two models using the test set
y_hat_full <- predict(model_full, newdata = df_reg_test)
y_hat_peers <- predict(model_peers, newdata = df_reg_test)
y_hat_indiv <- predict(model_indiv, newdata = df_reg_test)

# Compute MSE
MSE_full <- sum((df_reg_test$PUNT_GLOBAL - y_hat_full)**2) / nrow(df_reg_test)
MSE_peers <- sum((df_reg_test$PUNT_GLOBAL - y_hat_peers)**2) / nrow(df_reg_test)
MSE_indiv <- sum((df_reg_test$PUNT_GLOBAL - y_hat_indiv)**2) / nrow(df_reg_test)

MSE_full
MSE_peers
MSE_indiv
```

Either with R^2 or with test set we can see that the peers mean SES is as good (even better) than the individual SES.

```{r}
# Plot the covariates for the regressions
par(mfrow = c(1,2))

plot(df_reg$mean_ses, df_reg$PUNT_GLOBAL, pch = 16, cex = 0.5,
     main = "Scores vs peers SES",
     xlab = "peers SES", ylab = "Scores")
abline(model_peers, col = "red")

plot(df_reg$ESTU_INSE_INDIVIDUAL, df_reg$PUNT_GLOBAL, pch = 16, cex = 0.5,
     main = "Scores vs individual SES",
     xlab = "individual SES", ylab = "Scores")
abline(model_indiv, col = "red")
```


This is a significant result knowing that there are much less individual values for peer SES.


```{r}
library(sf)
library(ggplot2)

# Import a geojson or shapefile
map <- read_sf("Data/poblacion-upz-bogota.geojson") 
# potrebbero non essere sufficienti, perché ci sono scuole fuori dal confine della mappa
bogota.map <- ggplot(map) +
                geom_sf(fill = "white") +
                theme_void()

coord <- as.matrix(data_schools[, 3:4])

final_map <- bogota.map +
  geom_point(data = as.data.frame(coord), aes(x = X, y = Y, color= data_schools$indiv_SES), size = 0.15)+
  scale_color_gradient(low="red", high="green", name="Scores")+
  theme(legend.position="right", axis.text.x = element_text(size = 7))
final_map
```
